moved sampler utils; mostly ported line in; removed more junk
