--- scalaudio-buffer/src/main/scala/scalaudio/buffer/unitgen/LineIn.scala
@@ -1,15 +1,0 @@
-package scalaudio.buffer.unitgen
-
-import scalaudio.core.AudioContext
-import scalaudio.core.engine.Interleaver
-
-/**
-  * Created by johnmcgill on 12/20/15.
-  */
-case class LineIn()(implicit audioContext: AudioContext) extends UnitGen {
-  override def computeBuffer(params : Option[UnitParams] = None) = {
-    var buffer : Array[Double] = Array[Double](audioContext.config.framesPerBuffer * audioContext.config.nInChannels)
-    audioContext.audioInput.read(buffer)
-    internalBuffers = if (audioContext.config.nInChannels > 1) Interleaver.deinterleave(buffer, audioContext.config.nInChannels) else List(buffer)
-  }
-}
